+++ src/core/commands/index.ts
@@ -7,7 +7,7 @@ import "./internalHooksCommand.js";
 import "./memoryCommand.js";
 import "./mcpCommand.js";
 import "./workspaceCommand.js";
-import "./indexCommand.js";
+
 
 export { registry } from "./registry.js";
 export * from "./types.js";